fix: refactor cli version resolution to isolate configuration merging
